Rename LifeGame.tsx => index.tsx
