+++ docs/02advanced/01life-game/index.mdx
@@ -4,7 +4,7 @@ sidebar_position: 1
 
 import ViewSource from "@site/src/components/ViewSource/ViewSource";
 import Answer from "@site/src/components/Answer";
-import LifeGame from "@site/src/components/LifeGame/LifeGame";
+import LifeGame from "@site/src/components/LifeGame";
 
 # ライフゲーム
