+++ docs/03extras/04website/02docusaurus/01markdown/index.mdx
@@ -2,8 +2,8 @@
 sidebar_position: 1
 ---
 
-import LifeGame from "@site/src/components/LifeGame/LifeGame";
 import InteractiveCodeEditor from "@site/src/components/InteractiveCodeEditor";
+import LifeGame from "@site/src/components/LifeGame";
 export const defaultMathValue = `\
 $a^2 + b^2 = c^2$
 $$
