+++ src/components/LifeGame/index.tsx
